+++ server/feature/auth/authmiddleware/auth_middleware.go
@@ -9,12 +9,11 @@ import (
 	"github.com/sbondCo/Watcharr/config"
 	"github.com/sbondCo/Watcharr/database/entity"
 	"github.com/sbondCo/Watcharr/feature/auth/permission"
-	"gorm.io/gorm"
+	"github.com/sbondCo/Watcharr/feature/user/usermiddleware"
 )
 
 // Auth middleware
-// If db is passed, extra user info from the database will be fetched.
-func AuthRequired(db *gorm.DB, cfg *config.ServerConfig) gin.HandlerFunc {
+func AuthRequired(cfg *config.ServerConfig) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		slog.Debug("AuthRequired middleware hit")
 		atoken := c.GetHeader("Authorization")
@@ -47,23 +46,6 @@ func AuthRequired(db *gorm.DB, cfg *config.ServerConfig) gin.HandlerFunc {
 			slog.Debug("Token is valid", "claims", claims)
 			c.Set("userId", claims.UserID)
 			c.Set("userType", claims.Type)
-			// If db passed, get extra user info and set as variables in req context
-			if db != nil {
-				slog.Debug("AuthRequired: db passed.. getting extra user info")
-				dbUser := new(entity.User)
-				res := db.Where("id = ?", claims.UserID).Take(dbUser)
-				if res.Error != nil {
-					slog.Error("AuthRequired: Failed to select user from database", "error", res.Error)
-					c.AbortWithStatus(401)
-					return
-				}
-				slog.Debug("AuthRequired: fetched extra user info. Setting vars.")
-				c.Set("username", dbUser.Username)
-				c.Set("userPermissions", dbUser.Permissions)
-				if dbUser.Country != nil {
-					c.Set("userCountry", *dbUser.Country)
-				}
-			}
 			c.Next()
 		} else {
 			slog.Error("Token is **not** valid")
@@ -73,32 +55,24 @@ func AuthRequired(db *gorm.DB, cfg *config.ServerConfig) gin.HandlerFunc {
 	}
 }
 
-// Admin only middleware (use after AuthRequired with extra info!)
+// Admin only middleware (only use after AuthRequired & usermiddleware.WithUser!)
 func AdminRequired() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		userId := c.GetUint("userId")
-		perms := c.GetInt("userPermissions")
-		if permission.Has(perms, entity.PERM_ADMIN) {
-			slog.Debug("AdminRequired: User has permission to access admin only route", "user_id", userId)
-			c.Next()
-			return
-		}
-		slog.Info("AdminRequired: User denied permission to access admin only route", "user_id", userId)
-		c.AbortWithStatus(401)
-	}
+	return PermRequired(entity.PERM_ADMIN)
 }
 
-// Specific perm only middleware (use after AuthRequired with extra info!)
+// Specific perm only middleware (only use after AuthRequired & usermiddleware.WithUser!)
 func PermRequired(perm int) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		userId := c.GetUint("userId")
-		perms := c.GetInt("userPermissions")
-		if permission.Has(perms, perm) {
-			slog.Debug("PermRequired: User has permission to access perm only route", "user_id", userId, "required_perm", perm)
+		user := usermiddleware.UserFromContext(c)
+		if permission.Has(user.Permissions, perm) {
+			slog.Debug("PermRequired: User has permission!",
+				"user_id", userId, "required_perm", perm)
 			c.Next()
 			return
 		}
-		slog.Info("PermRequired: User denied permission to access perm only route", "user_id", userId, "required_perm", perm)
+		slog.Info("PermRequired: User denied permission!",
+			"user_id", userId, "required_perm", perm)
 		c.AbortWithStatus(401)
 	}
 }